backend/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
prophet
joblib
firebase-admin
google-genai
gunicorn
python-dotenv
ortools